# docs/reference/host-api-v1.openapi.yaml
openapi: 3.0.3
info:
  title: LIGHT-BELT Host API
  version: "1.0"
  description: |
    LIGHT-BELT APP 接入接口。APP 通过 HTTPS REST JSON 与 WSS WebSocket JSON 接入 RK3588 Host Service。

    RK3588 IP 已确定为 192.168.31.236。当前 LIGHT-BELT Host Service 尚未部署到 RK3588，本 OpenAPI 用于 APP 开发、Mock 和接口冻结。

    HTTPS Base URL: https://192.168.31.236:8443/api/v1

    WebSocket URL: wss://192.168.31.236:8443/ws

    HTTPS Base URL 和 WebSocket URL 是部署后的固定对接地址。

    Certificate Fingerprint: 预生成证书指纹 0A:14:3C:FA:A3:FD:4F:CA:54:94:F4:7E:FE:EE:68:33:F9:BC:5E:29:91:76:DF:66:7C:95:A5:F8:E7:44:9D:4D

    Certificate SAN: DNS: light-belt-rk3588.local, IP: 192.168.31.236

    实际联调以部署到 RK3588 的 Host Service 使用的证书为准。
servers:
  - url: https://192.168.31.236:8443/api/v1
    description: RK3588 Host Service 部署后固定对接地址
...
security:
  - bearerAuth: []
paths:
...
  /auth/pair:
    post:
      tags: [Auth]
      summary: 使用配对码换取访问令牌
      security: []
      parameters:
        - $ref: "#/components/parameters/XRequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PairRequest"
            example:
              pairing_code: "123456"
              client_id: tablet-main-room-001
              client_name: Main Room Tablet
              client_type: tablet
              app_version: 1.0.0
      responses:
        "200":
          description: 配对成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
              example:
                ok: true
                request_id: req-pair-001
                data:
                  access_token: eyJhbGciOiJIUzI1NiIs...
                  refresh_token: rt_7d4f3b8f...
                  token_type: Bearer
                  expires_in: 3600
                  scope:
                    - state:read
                    - playback:write
                    - lights:write
                    - effects:write
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
  /auth/refresh:
    post:
      tags: [Auth]
      summary: 刷新访问令牌
      security: []
      parameters:
        - $ref: "#/components/parameters/XRequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
            example:
              refresh_token: rt_7d4f3b8f...
      responses:
        "200":
          description: 刷新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
              example:
                ok: true
                request_id: req-refresh-001
                data:
                  access_token: eyJhbGciOiJIUzI1NiIs...
                  refresh_token: rt_90a2c1d5...
                  token_type: Bearer
                  expires_in: 3600
                  scope:
                    - state:read
                    - playback:write
                    - lights:write
                    - effects:write
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
  /session/ws-ticket:
    post:
      tags: [Session]
      summary: 申请 WebSocket 连接票据
      parameters:
        - $ref: "#/components/parameters/XRequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WsTicketRequest"
            example:
              subscribe:
                - runtime.state
                - playback.progress
                - device.status
                - error.event
                - heartbeat
      responses:
        "200":
          description: ticket 已生成
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WsTicketResponse"
              example:
                ok: true
                request_id: req-ticket-001
                data:
                  ws_ticket: wst_3f610c1a...
                  session_id: sess_20260709_0001
                  expires_in: 60
                  ws_url: wss://192.168.31.236:8443/ws?ticket=wst_3f610c1a...
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: access_token
  parameters:
    XRequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
      description: APP 生成的请求 ID。
  responses:
    ErrorResponse:
      description: 错误响应
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            ok: false
            request_id: req-20260709-0001
            error:
              code: INVALID_ARGUMENT
              message: target_id is required
              details:
                field: target_id
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [ok, request_id, data]
      properties:
        ok:
          type: boolean
          enum: [true]
        request_id:
          type: string
        data:
          type: object
    ErrorEnvelope:
      type: object
      required: [ok, request_id, error]
      properties:
        ok:
          type: boolean
          enum: [false]
        request_id:
          type: string
        error:
          $ref: "#/components/schemas/ErrorObject"
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorCode:
      type: string
      enum:
        - INVALID_ARGUMENT
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - TOKEN_EXPIRED
        - PAIRING_CODE_INVALID
        - WS_TICKET_INVALID
        - DEVICE_OFFLINE
        - SHOW_NOT_LOADED
        - PLAYBACK_NOT_READY
        - MPV_UNAVAILABLE
        - SCENE_LIMIT_EXCEEDED
        - INTERNAL_ERROR
    ClientType:
      type: string
      enum: [tablet, phone, debug]
...
    WsMessageType:
      type: string
      enum:
        - session.connected
        - runtime.state
        - playback.progress
        - device.status
        - error.event
        - heartbeat
...
    PairRequest:
      type: object
      required: [pairing_code, client_id, client_name, client_type, app_version]
      properties:
        pairing_code:
          type: string
          minLength: 1
        client_id:
          type: string
          minLength: 1
        client_name:
          type: string
          minLength: 1
        client_type:
          $ref: "#/components/schemas/ClientType"
        app_version:
          type: string
          minLength: 1
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          minLength: 1
    TokenData:
      type: object
      required: [access_token, refresh_token, token_type, expires_in, scope]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: number
          minimum: 0
        scope:
          type: array
          items:
            type: string
    AuthTokenResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/TokenData"
    WsTicketRequest:
      type: object
      required: [subscribe]
      properties:
        subscribe:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/WsMessageType"
    WsTicketData:
      type: object
      required: [ws_ticket, session_id, expires_in, ws_url]
      properties:
        ws_ticket:
          type: string
        session_id:
          type: string
        expires_in:
          type: number
          minimum: 0
        ws_url:
          type: string
          format: uri
    WsTicketResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/WsTicketData"
...